Split `LocationService` into separate files
Generally, there’s a separate file for each top-level type. No
functional changes, other than some access control modifiers changing
from `private` to `internal` to allow the separation.
