+++ ELLocation/LocationUpdateService.swift
@@ -1,0 +1,47 @@
+//
+//  LocationUpdateService.swift
+//  ELLocation
+//
+//  Created by Alex Johnson on 3/16/16.
+//  Copyright © 2016 WalmartLabs. All rights reserved.
+//
+
+// A protocol for a type that wants to provide location updates.
+public protocol LocationUpdateProvider {
+    func registerListener(listener: AnyObject, request: LocationUpdateRequest) -> NSError?
+    func deregisterListener(listener: AnyObject)
+}
+
+// The interface for requesting location updates. Listeners can register to be informed of location updates
+// They can request to be deregistered or will be deregistered automatically when they are dealloced.
+public struct LocationUpdateService: LocationUpdateProvider {
+    let locationProvider: LocationUpdateProvider
+
+    /**
+     Registers a listener to receive location updates as per the parameters defined in the request.
+
+     - parameter listener: The listener to register.
+     - parameter request: The parameters of the request.
+     - returns: An optional error that could happen when registering. See `ELLocationError`.
+     */
+    public func registerListener(listener: AnyObject, request: LocationUpdateRequest) -> NSError? {
+        return locationProvider.registerListener(listener, request: request)
+    }
+
+    /**
+     Deregisters a listener from receiving any more location updates.
+
+     - parameter listener: The listener to deregister.
+     */
+    public func deregisterListener(listener: AnyObject) {
+        locationProvider.deregisterListener(listener)
+    }
+
+    public init() {
+        self.init(locationProvider: LocationManager.shared)
+    }
+
+    init(locationProvider: LocationUpdateProvider) {
+        self.locationProvider = locationProvider
+    }
+}
